requirements.txt
python-telegram-bot[job-queue]==22.6
openai>=1.0.0
requests>=2.31.0
pytz>=2024.1
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
